Loopback/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-anymail==13.0
django-celery-beat==2.8.0
django-crispy-forms==2.4
django-filter==25.1
django-sendgrid-v5==1.2.4
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.3
Markdown==3.8
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
sendgrid==6.11.0
sendgrid-django==4.2.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
